Death of actors
    -   Tried to make sure dead actors are removed and won't attack back.
        Keep investigating this.

    1. In GameLoop.hpp:
       After each action(), loop through the ACTORS and see if an Actor is_dead.
       If so, remove it.

    2. Created failing test case of 1.

    3. Fixed said test case and found another bug in the meantime.
       Dead enemies should always have zero hp. Not overflow. :D

+++ test.cpp
@@ -102,7 +102,7 @@ int main(int argc, char const *argv[]) {
         }
         tezt::ae(s1 - 1, ACTORS.size(), "One less actor than before.");
 
-        tezt::ae(0, rabbit->current_health, "Dead enemies have zero hp left. Nothing else.");
+        tezt::ae((uint32_t) 0, rabbit->current_health, "Dead enemies have zero hp left. Nothing else.");
     });
 
     return tezt::run();
